feat: added unique labeling
